backend/pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "A configurable Retrieval-Augmented Generation (RAG) system with pluggable Embedding, VectorStore, BM25, Rerank providers and a Web control plane."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "jieba>=0.42",
    "numpy>=1.26",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py310"
line-length = 120
